feat #139: create objection endpoint

+++ api/Domain/Entities/Objection/Models/Objection.cs
@@ -4,8 +4,8 @@ namespace Api.Domain.Models;
 
 public partial class Objection : IEntity
 {
-    public required double ClaimedAptitude { get; set; }
-    public required double OfficialAptitude { get; set; }
+    public required short ClaimedAptitude { get; set; }
+    public required short OfficialAptitude { get; set; }
     public required DateOnly ObjectedAt { get; set; }
     public bool IsAccepted { get; set; } = false;
     public bool IsActive { get; set; } = true;
